Update project structure
 - move tests/ to src/tests/
 - replace relative imports
