+++ src/app/models/hero.py
@@ -1,0 +1,8 @@
+from sqlmodel import SQLModel, Field
+
+
+class Hero(SQLModel, table=True):
+    id: int | None = Field(default=None, primary_key=True)
+    name: str = Field(index=True)
+    age: int | None = Field(default=None, index=True)
+    secret_name: str
