+++ src/app/models/team.py
@@ -1,0 +1,7 @@
+from sqlmodel import Field, SQLModel 
+
+
+class Team(SQLModel, table=True):
+    id: int | None = Field(default=None, primary_key=True)
+    name: str = Field(index=True)
+    headquarters: str
